--- processor/src/main/java/org/frc5572/robotools/BoolVisitor.java
@@ -1,13 +1,0 @@
-package org.frc5572.robotools;
-
-import javax.lang.model.util.SimpleAnnotationValueVisitor8;
-
-/** Find all uses of a string in annotation values */
-public class BoolVisitor extends SimpleAnnotationValueVisitor8<Boolean, Void> {
-
-    @Override
-    public Boolean visitBoolean(boolean arg0, Void arg1) {
-        return arg0;
-    }
-
-}
